backend/requirements.txt
# FastAPI et dépendances principales
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Base de données
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# HTTP et API externes
httpx==0.25.2
requests==2.31.0

# OpenAI
openai==1.3.7

# Utilitaires
python-dotenv==1.0.0
python-multipart==0.0.6
jinja2==3.1.2

# Développement et tests
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0

# CORS pour le frontend
python-cors==1.7.0
